+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-21T11:04:04.625Z")
+    created: Optional[datetime] = parse_datetime("2025-05-21T11:42:40.449Z")
     r"""The date and time the resource was created"""