+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVS87NPPR7E6VY09JWJCYARD"
+        "01JVSAEB7WK3R5DHKBCX53PHTZ"
     )
     r"""The id of the resource"""